restructure utils.timing
